requirements.txt
boto3==1.23.1
botocore==1.26.1
colorama==0.4.6
cryptography==3.4.8
ipykernel==6.17.1
ipython==8.7.0
jupyter_client==7.4.7
jupyter_core==5.1.0
numpy==1.24.1
oauthlib==3.2.0
openpyxl==3.0.9
pandas==1.4.1
prompt-toolkit==3.0.33
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==7.0.0
PyJWT==2.3.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.20.0
PyYAML==5.4.1
pyzmq==24.0.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.46
tornado==6.2
urllib3==1.26.9
requests==2.28.1
stripe==5.0.0
